Fix issue that would cause non-responsiveness
The program would throw the error: "input hex char count for 10000 is greater then the output"

"10000" is a hexadecimal representation of the transaction id (tid)

The transaction id is 2 byte a fixed length value. The largest it can be is 65535 (decimal) or FFFF (hex).

The transaction id (tid) is supposed to reset once it grows larger then the packetBufferLength

The lastTid mistakenly had "this" in front of it. Causing the tid to grow until it reached 65536, which cannot fit in the 2 byte space defined by modbus.